requirements.txt
pandas
numpy
opencv-python
flask
keras
tensorflow
pillow